Update BGBuddy_v1.ino to v1.3
Updated to v1.3 which includes improvements to WiFi connectivity and a change to how the Nightscout API was being called, allowing it to handle a larger number of configurations. Also included in this update is a second configuration for the IdeaSpark ESP8266 w/ OLED package; a user reported a package where the data and clock pins are reversed and so there will be both an "SDA12" [sda on pin 12] and "SDA14" [sda on pin 14] version of the binary.